+++ libs/utils/Emitter.lua
@@ -1,8 +1,11 @@
+local Listener = require('./Listener')
+local Iterable = require('./Iterable')
+
 local class = require('../class')
 local typing = require('../typing')
 local helpers = require('../helpers')
 
-local wrap, yield, running = coroutine.wrap, coroutine.yield, coroutine.running
+local yield, running = coroutine.yield, coroutine.running
 local insert, remove = table.insert, table.remove
 local setTimeout, clearTimer = helpers.setTimeout, helpers.clearTimer
 local checkType = typing.checkType
@@ -32,43 +35,34 @@ local function clean(listeners)
 	listeners.marked = nil
 end
 
+local once = {}
+
 function Emitter:__init()
 	self._listeners = setmetatable({}, meta)
 end
 
-function Emitter:on(name, fn, err)
-	insert(self._listeners[checkType('string', name)], {
-		fn = checkType('function', fn),
-		err = err and checkType('function', err),
-	})
-	return fn
+function Emitter:on(eventName, callback, errorHandler)
+	local listener = Listener(self, eventName, callback, errorHandler)
+	insert(self._listeners[listener.eventName], listener)
+	return listener
 end
 
-function Emitter:once(name, fn, err)
-	insert(self._listeners[checkType('string', name)], {
-		fn = checkType('function', fn),
-		err = err and checkType('function', err),
-		once = true
-	})
-	return fn
+function Emitter:once(eventName, callback, errorHandler)
+	local listener = Listener(self, eventName, callback, errorHandler)
+	insert(self._listeners[listener.eventName], listener)
+	once[listener] = true
+	return listener
 end
 
-function Emitter:emit(name, ...)
-	local listeners = self._listeners[checkType('string', name)]
+function Emitter:emit(eventName, ...)
+	local listeners = self._listeners[checkType('string', eventName)]
 	for i = 1, #listeners do
 		local listener = listeners[i]
 		if listener then
-			if listener.once then
+			if once[listener] then
 				mark(listeners, i)
 			end
-			if listener.err then
-				local success, err = pcall(wrap(listener.fn), ...)
-				if not success then
-					wrap(listener.err)(err, ...)
-				end
-			else
-				wrap(listener.fn)(...)
-			end
+			listener:fire(...)
 		end
 	end
 	if listeners.marked then
@@ -76,43 +70,31 @@ function Emitter:emit(name, ...)
 	end
 end
 
-function Emitter:getListeners(name)
-	local listeners = self._listeners[checkType('string', name)]
-	local i = 0
-	return function()
-		while i < #listeners do
-			i = i + 1
-			if listeners[i] then
-				return listeners[i].fn
-			end
+function Emitter:getListeners(eventName)
+	local listeners = self._listeners[checkType('string', eventName)]
+	local new = {}
+	for _, v in ipairs(listeners) do
+		if v then
+			insert(new, v)
 		end
 	end
+	return Iterable(new)
 end
 
-function Emitter:getListenerCount(name)
-	local listeners = self._listeners[checkType('string', name)]
-	local n = 0
-	for _, listener in ipairs(listeners) do
-		if listener then
-			n = n + 1
+function Emitter:removeListener(eventName, listener)
+	local listeners = self._listeners[checkType('string', eventName)]
+	for i, v in ipairs(listeners) do
+		if v == listener then
+			listeners[i] = false
+			return true
 		end
 	end
-	return n
+	return false
 end
 
-function Emitter:removeListener(name, fn)
-	local listeners = self._listeners[checkType('string', name)]
-	for i, listener in ipairs(listeners) do
-		if listener and listener.fn == fn then
-			mark(listeners, i)
-			break
-		end
-	end
-end
-
-function Emitter:removeAllListeners(name)
-	if name then
-		self._listeners[checkType('string', name)] = nil
+function Emitter:removeAllListeners(eventName)
+	if eventName then
+		self._listeners[checkType('string', eventName)] = nil
 	else
 		for k in pairs(self._listeners) do
 			self._listeners[k] = nil
@@ -120,12 +102,12 @@ function Emitter:removeAllListeners(name)
 	end
 end
 
-function Emitter:waitFor(name, timeout, predicate)
+function Emitter:waitFor(eventName, timeout, predicate)
 
-	name = checkType('string', name)
+	eventName = checkType('string', eventName)
 	predicate = predicate and checkType('function', predicate)
 
-	local t, fn
+	local t, listener
 	local thread = running()
 
 	local function complete(success, ...)
@@ -133,14 +115,14 @@ function Emitter:waitFor(name, timeout, predicate)
 			clearTimer(t)
 			t = nil
 		end
-		if fn then
-			self:removeListener(name, fn)
-			fn = nil
+		if listener then
+			self:removeListener(eventName, listener)
+			listener = nil
 			return assertResume(thread, success, ...)
 		end
 	end
 
-	fn = self:on(name, function(...)
+	listener = self:on(eventName, function(...)
 		if not predicate or predicate(...) then
 			return complete(true, ...)
 		end